More of the complex number function unit tests refactored
