+++ tests/PhpSpreadsheetTests/Calculation/EngineeringTest.php
@@ -51,146 +51,6 @@ class EngineeringTest extends TestCase
     }
 
     /**
-     * @dataProvider providerIMCOS
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMCOS($expectedResult, $value)
-    {
-        $result = Engineering::IMCOS($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMCOS()
-    {
-        return require 'data/Calculation/Engineering/IMCOS.php';
-    }
-
-    /**
-     * @dataProvider providerIMCOSH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMCOSH($expectedResult, $value)
-    {
-        $result = Engineering::IMCOSH($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMCOSH()
-    {
-        return require 'data/Calculation/Engineering/IMCOSH.php';
-    }
-
-    /**
-     * @dataProvider providerIMCOT
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMCOT($expectedResult, $value)
-    {
-        $result = Engineering::IMCOT($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMCOT()
-    {
-        return require 'data/Calculation/Engineering/IMCOT.php';
-    }
-
-    /**
-     * @dataProvider providerIMCSC
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMCSC($expectedResult, $value)
-    {
-        $result = Engineering::IMCSC($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMCSC()
-    {
-        return require 'data/Calculation/Engineering/IMCSC.php';
-    }
-
-    /**
-     * @dataProvider providerIMCSCH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMCSCH($expectedResult, $value)
-    {
-        $result = Engineering::IMCSCH($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMCSCH()
-    {
-        return require 'data/Calculation/Engineering/IMCSCH.php';
-    }
-
-    /**
-     * @dataProvider providerIMSEC
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMSEC($expectedResult, $value)
-    {
-        $result = Engineering::IMSEC($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMSEC()
-    {
-        return require 'data/Calculation/Engineering/IMSEC.php';
-    }
-
-    /**
-     * @dataProvider providerIMSECH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMSECH($expectedResult, $value)
-    {
-        $result = Engineering::IMSECH($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMSECH()
-    {
-        return require 'data/Calculation/Engineering/IMSECH.php';
-    }
-
-    /**
      * @dataProvider providerIMDIV
      *
      * @param mixed $expectedResult
@@ -207,86 +67,6 @@ class EngineeringTest extends TestCase
     public function providerIMDIV()
     {
         return require 'data/Calculation/Engineering/IMDIV.php';
-    }
-
-    /**
-     * @dataProvider providerIMEXP
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMEXP($expectedResult, $value)
-    {
-        $result = Engineering::IMEXP($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMEXP()
-    {
-        return require 'data/Calculation/Engineering/IMEXP.php';
-    }
-
-    /**
-     * @dataProvider providerIMLN
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMLN($expectedResult, $value)
-    {
-        $result = Engineering::IMLN($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMLN()
-    {
-        return require 'data/Calculation/Engineering/IMLN.php';
-    }
-
-    /**
-     * @dataProvider providerIMLOG2
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMLOG2($expectedResult, $value)
-    {
-        $result = Engineering::IMLOG2($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMLOG2()
-    {
-        return require 'data/Calculation/Engineering/IMLOG2.php';
-    }
-
-    /**
-     * @dataProvider providerIMLOG10
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMLOG10($expectedResult, $value)
-    {
-        $result = Engineering::IMLOG10($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMLOG10()
-    {
-        return require 'data/Calculation/Engineering/IMLOG10.php';
     }
 
     /**
@@ -325,86 +105,6 @@ class EngineeringTest extends TestCase
     public function providerIMPRODUCT()
     {
         return require 'data/Calculation/Engineering/IMPRODUCT.php';
-    }
-
-    /**
-     * @dataProvider providerIMSIN
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMSIN($expectedResult, $value)
-    {
-        $result = Engineering::IMSIN($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMSIN()
-    {
-        return require 'data/Calculation/Engineering/IMSIN.php';
-    }
-
-    /**
-     * @dataProvider providerIMSINH
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMSINH($expectedResult, $value)
-    {
-        $result = Engineering::IMSINH($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMSINH()
-    {
-        return require 'data/Calculation/Engineering/IMSINH.php';
-    }
-
-    /**
-     * @dataProvider providerIMTAN
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMTAN($expectedResult, $value)
-    {
-        $result = Engineering::IMTAN($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMTAN()
-    {
-        return require 'data/Calculation/Engineering/IMTAN.php';
-    }
-
-    /**
-     * @dataProvider providerIMSQRT
-     *
-     * @param mixed $expectedResult
-     * @param mixed $value
-     */
-    public function testIMSQRT($expectedResult, $value)
-    {
-        $result = Engineering::IMSQRT($value);
-        self::assertTrue(
-            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
-            $this->complexAssert->getErrorMessage()
-        );
-    }
-
-    public function providerIMSQRT()
-    {
-        return require 'data/Calculation/Engineering/IMSQRT.php';
     }
 
     /**
