+++ tests/PhpSpreadsheetTests/Calculation/Functions/Engineering/ImCotTest.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Engineering;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Engineering;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheetTests\Custom\ComplexAssert;
+use PHPUnit\Framework\TestCase;
+
+class ImCotTest extends TestCase
+{
+    const COMPLEX_PRECISION = 1E-8;
+
+    /**
+     * @var ComplexAssert
+     */
+    protected $complexAssert;
+
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        $this->complexAssert = new ComplexAssert();
+    }
+
+    public function tearDown()
+    {
+        $this->complexAssert = null;
+    }
+
+    /**
+     * @dataProvider providerIMCOT
+     *
+     * @param mixed $expectedResult
+     * @param mixed $value
+     */
+    public function testIMCOT($expectedResult, $value)
+    {
+        $result = Engineering::IMCOT($value);
+        $this->assertTrue(
+            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
+            $this->complexAssert->getErrorMessage()
+        );
+    }
+
+    public function providerIMCOT()
+    {
+        return require 'data/Calculation/Engineering/IMCOT.php';
+    }
+}
