+++ tests/PhpSpreadsheetTests/Calculation/Functions/Engineering/ImCschTest.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Engineering;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Engineering;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheetTests\Custom\ComplexAssert;
+use PHPUnit\Framework\TestCase;
+
+class ImCschTest extends TestCase
+{
+    const COMPLEX_PRECISION = 1E-8;
+
+    /**
+     * @var ComplexAssert
+     */
+    protected $complexAssert;
+
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        $this->complexAssert = new ComplexAssert();
+    }
+
+    public function tearDown()
+    {
+        $this->complexAssert = null;
+    }
+
+    /**
+     * @dataProvider providerIMCSCH
+     *
+     * @param mixed $expectedResult
+     * @param mixed $value
+     */
+    public function testIMCSCH($expectedResult, $value)
+    {
+        $result = Engineering::IMCSCH($value);
+        $this->assertTrue(
+            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
+            $this->complexAssert->getErrorMessage()
+        );
+    }
+
+    public function providerIMCSCH()
+    {
+        return require 'data/Calculation/Engineering/IMCSCH.php';
+    }
+}
