+++ tests/PhpSpreadsheetTests/Calculation/Functions/Engineering/ImLnTest.php
@@ -1,0 +1,49 @@
+<?php
+
+namespace PhpOffice\PhpSpreadsheetTests\Calculation\Functions\Engineering;
+
+use PhpOffice\PhpSpreadsheet\Calculation\Engineering;
+use PhpOffice\PhpSpreadsheet\Calculation\Functions;
+use PhpOffice\PhpSpreadsheetTests\Custom\ComplexAssert;
+use PHPUnit\Framework\TestCase;
+
+class ImLnTest extends TestCase
+{
+    const COMPLEX_PRECISION = 1E-8;
+
+    /**
+     * @var ComplexAssert
+     */
+    protected $complexAssert;
+
+    public function setUp()
+    {
+        Functions::setCompatibilityMode(Functions::COMPATIBILITY_EXCEL);
+        $this->complexAssert = new ComplexAssert();
+    }
+
+    public function tearDown()
+    {
+        $this->complexAssert = null;
+    }
+
+    /**
+     * @dataProvider providerIMLN
+     *
+     * @param mixed $expectedResult
+     * @param mixed $value
+     */
+    public function testIMLN($expectedResult, $value)
+    {
+        $result = Engineering::IMLN($value);
+        $this->assertTrue(
+            $this->complexAssert->assertComplexEquals($expectedResult, $result, self::COMPLEX_PRECISION),
+            $this->complexAssert->getErrorMessage()
+        );
+    }
+
+    public function providerIMLN()
+    {
+        return require 'data/Calculation/Engineering/IMLN.php';
+    }
+}
